requirements.txt
streamlit==1.40.0
plotly==5.24.0
pandas==2.2.3
sqlalchemy==2.0.36
openpyxl==3.1.5
python-dotenv==1.0.1
psycopg2-binary==2.9.9
